+++ forui/CHANGELOG.md
@@ -1,3 +1,39 @@
+## 0.14.0 (Next)
+
+### `FButton`
+* Add `FButton.onSecondaryPress`.
+* Add `FButton.onSecondaryLongPress`.
+
+
+### `FHeaderAction`
+* Add `FHeaderAction.onSecondaryPress`.
+* Add `FHeaderAction.onSecondaryLongPress`.
+* Add `FHeaderAction.actions`.
+* Add `FHeaderAction.shortcuts`.
+
+
+### `FSelect`
+* Fix `FSelect` flickering when selecting an item on desktop.
+* Fix `FSelect` not scrolling to selected item when opened if the item is really far down the list.
+* Fix `FSelect` layout shifting when scrolling through a long list of items.
+
+
+### `FTappable`
+* Add `FTappable.onSecondaryPress`.
+* Add `FTappable.onSecondaryLongPress`.
+
+
+### `FTile`
+* Add `FTile.onSecondaryPress`.
+* Add `FTile.onSecondaryLongPress`.
+
+
+### Others
+* **Breaking** Change `FPersistentSheetController.shown` to `FPersistentSheetController.status`.
+* **Breaking** Change `FPopoverController.shown` to `FPopoverController.status`.
+* **Breaking** Change `FTooltipController.shown` to `FTooltipController.status`.
+
+
 ## 0.13.1
 
 * Fix `FSelectTile` mixing in `FItemMixin` instead of `FTileMixin`.